+++ NEWS.md
@@ -1,9 +1,8 @@
 # scorecard 0.1.3
 
-* Modified the format of printing message.
+* Modified the format of printing message and added condition functions.
 * Added the split_df function which split a dataframe into two.
 * Reorder the binning information. Move the missing to the first binning.
-* Added condition functions.
 
 # scorecard 0.1.2